FEATURE: consolidates spirits data/types into one file
