+++ libs/spirit-islander/util/src/lib/data/spirits.ts
@@ -1,4 +1,37 @@
-import type { Spirit } from '../types/game/spirits';
+import { ExpansionOption } from '../types/game/expansions';
+
+export type Spirit = ExpansionOption<SpiritName>;
+
+export type SpiritName =
+  | 'A Spread of Rampant Green'
+  | 'Bringer of Dreams and Nightmares'
+  | 'Devouring Teeth Lurk Underfoot'
+  | 'Downpour Drenches the World'
+  | 'Eyes Watch From the Trees'
+  | 'Fathomless Mud of the Swamp'
+  | 'Finder of Paths Unseen'
+  | 'Fractured Days Split the Sky'
+  | 'Grinning Trickster Stirs Up Trouble'
+  | 'Heart of the Wildfire'
+  | 'Keeper of the Forbidden Wilds'
+  | "Lightning's Swift Strike"
+  | 'Lure of the Deep Wilderness'
+  | 'Many Minds Move as One'
+  | "Ocean's Hungry Grasp"
+  | 'Rising Heat of Stone and Sand'
+  | 'River Surges in Sunlight'
+  | 'Serpent Slumbering Beneath the Island'
+  | 'Shadows Flicker Like Flame'
+  | 'Sharp Fangs Behind the Leaves'
+  | 'Shifting Memory of Ages'
+  | 'Shroud of Silent Mist'
+  | 'Sun-Bright Whirlwind'
+  | 'Starlight Seeks Its Form'
+  | "Stone's Unyielding Defiance"
+  | 'Thunderspeaker'
+  | 'Vengeance as a Burning Plague'
+  | 'Vital Strength of the Earth'
+  | 'Volcano Looming High';
 
 export const SPIRITS: readonly Spirit[] = [
   { name: 'A Spread of Rampant Green' },
@@ -30,4 +63,4 @@ export const SPIRITS: readonly Spirit[] = [
   { name: 'Vengeance as a Burning Plague', expansion: 'Jagged Earth' },
   { name: 'Vital Strength of the Earth' },
   { name: 'Volcano Looming High', expansion: 'Jagged Earth' },
-] as const;
+];
