+++ libs/spirit-islander/util/src/lib/types/config.interface.ts
@@ -3,7 +3,7 @@ import type { Difficulty } from './game/difficulty';
 import type { MapName } from './game/maps';
 import type { BalancedBoardName } from './game/board';
 import type { ExpansionName } from './game/expansions';
-import type { SpiritName } from './game/spirits';
+import type { SpiritName } from '../data/spirits';
 import type { AdversaryLevelId } from './game/adversaries';
 import type { ScenarioName } from './game/scenarios';
